refactor: Moved Downloaders to the Downloaders folder
